Refactor to comply with new CONTRIBUTING.md
