+++ bigbio/hub/hub_repos/bronco/README.md
@@ -1,0 +1,47 @@
+---
+language:
+  - de
+bigbio_language:
+  - German
+multilinguality: monolingual
+pretty_name: BRONCO150
+homepage: https://www2.informatik.hu-berlin.de/~leser/bronco/index.html
+bigbio_pubmed: false
+bigbio_public: true
+bigbio_tasks:
+  - NAMED_ENTITY_RECOGNITION
+  - NAMED_ENTITY_DISAMBIGUATION
+---
+
+
+# Dataset Card for BRONCO150
+
+## Dataset Description
+
+- **Homepage:** https://www2.informatik.hu-berlin.de/~leser/bronco/index.html
+- **Pubmed:** False
+- **Public:** True
+- **Tasks:** NER, NED
+
+
+BRONCO150 is a corpus containing selected sentences of 150 German discharge summaries of cancer patients (hepatocelluar carcinoma or melanoma) treated at Charite Universitaetsmedizin Berlin or Universitaetsklinikum Tuebingen. All discharge summaries were manually anonymized. The original documents were scrambled at the sentence level to make reconstruction of individual reports impossible.
+
+
+## Citation Information
+
+```
+@article{10.1093/jamiaopen/ooab025,
+    author = {Kittner, Madeleine and Lamping, Mario and Rieke, Damian T and Götze, Julian and Bajwa, Bariya and Jelas, Ivan and Rüter, Gina and Hautow, Hanjo and Sänger, Mario and Habibi, Maryam and Zettwitz, Marit and Bortoli, Till de and Ostermann, Leonie and Ševa, Jurica and Starlinger, Johannes and Kohlbacher, Oliver and Malek, Nisar P and Keilholz, Ulrich and Leser, Ulf},
+    title = "{Annotation and initial evaluation of a large annotated German oncological corpus}",
+    journal = {JAMIA Open},
+    volume = {4},
+    number = {2},
+    year = {2021},
+    month = {04},
+    issn = {2574-2531},
+    doi = {10.1093/jamiaopen/ooab025},
+    url = {https://doi.org/10.1093/jamiaopen/ooab025},
+    note = {ooab025},
+    eprint = {https://academic.oup.com/jamiaopen/article-pdf/4/2/ooab025/38830128/ooab025.pdf},
+}
+```
